third_party/blink/web_tests/svg/animations/animate-color-rgba-calcMode-discrete-expected.txt

SVG 1.1 dynamic animation tests

Test calcMode discrete with from-to animation on colors with alpha channel. You should see a green 100x100 rect and only PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS colors[0] is 0
PASS colors[1] is 255
PASS colors[2] is 255
PASS colors[3] is 0.4
PASS colors[0] is 255
PASS colors[1] is 0
PASS colors[2] is 0
PASS colors[3] is 0.6
PASS colors[0] is 255
PASS colors[1] is 0
PASS colors[2] is 0
PASS colors[3] is 0.6
PASS colors[0] is 0
PASS colors[1] is 255
PASS colors[2] is 255
PASS colors[3] is 0.8
PASS colors[0] is 0
PASS colors[1] is 255
PASS colors[2] is 255
PASS colors[3] is 0.8
PASS colors[0] is 0
PASS colors[1] is 255
PASS colors[2] is 255
PASS colors[3] is 0.4
PASS successfullyParsed is true

TEST COMPLETE
